+++ CHANGELOG.md
@@ -6,9 +6,18 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 The following changes have been implemented but not released yet:
 
+### New features
+
+- `redirectToRequestor` has been added to the `./manage` module to help access management
+  apps redirect their users to the clients which requested access to a Resource.
+
 ### Bugfixes
 
 - The JSON-LD context URL has changed from https://vc.inrupt.com/credentials/v1 to https://vc.inrupt.com/credentials/v1.
+- `redirectToAccessManagementUi` now returns a promise that never resolves. This
+  prevents from `await`ing on it, have it resolve, and start running code in the undeterministic
+  time between the moment when `window.location.href` is set (or the redirect callback
+  is called) and the moment the redirection actually happens.
 
 ### Deprecation